+++ routes/web.php
@@ -31,34 +31,37 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [JobController::class, 'index']);
 
 // Show/Get create form 
-Route::get("/jobs/create", [JobController::class, 'create']);
+Route::get("/jobs/create", [JobController::class, 'create'])->middleware('auth');
+
+// Manage Jobs for a user
+Route::get("/jobs/manage", [JobController::class, 'manage'])->middleware('auth');
 
 // Store/Create a new job
-Route::post("/jobs", [JobController::class, 'store']);
+Route::post("/jobs", [JobController::class, 'store'])->middleware('auth');
 
 // Show edit form
-Route::get("/jobs/{id}/edit", [JobController::class, 'edit']);
+Route::get("/jobs/{id}/edit", [JobController::class, 'edit'])->middleware('auth');
 
 // Update job data
-Route::put("/jobs/{id}", [JobController::class, 'update']);
+Route::put("/jobs/{id}", [JobController::class, 'update'])->middleware('auth');
 
 // Delete job data
-Route::delete("/jobs/{id}", [JobController::class, 'destroy']);
+Route::delete("/jobs/{id}", [JobController::class, 'destroy'])->middleware('auth');
 
 // Find one job
 Route::get("/jobs/{id}", [JobController::class, 'show']);
 
 // Show Register form
-Route::get("/signup", [UserController::class, 'signup']);
+Route::get("/signup", [UserController::class, 'signup'])->middleware('guest');
 
 // Store/Register a new user
 Route::post("/users", [UserController::class, 'store']);
 
 // Logout
-Route::post("/logout", [UserController::class, 'logout']);
+Route::post("/logout", [UserController::class, 'logout'])->middleware('auth');
 
 // Show Login form
-Route::get("/login", [UserController::class, 'login']);
+Route::get("/login", [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Handle Login form data
 Route::post("/users/authenticate", [UserController::class, 'authenticate']);
